adjust assimp to spring needs

+++ contrib/poly2tri/poly2tri/sweep/sweep.h
@@ -174,12 +174,12 @@ private:
    * @param node - middle node
    * @return the angle between 3 front nodes
    */
-  double HoleAngle(Node& node);
+  float HoleAngle(Node& node);
 
   /**
    * The basin angle is decided against the horizontal line [1,0]
    */
-  double BasinAngle(Node& node);
+  float BasinAngle(Node& node);
 
   /**
    * Fills a basin that has formed on the Advancing Front to the right
